finished backend for presentation

+++ src/main/java/uvt/tw/conferencemanagementsystem/app/conference/service/ConferenceService.java
@@ -1,15 +1,22 @@
 package uvt.tw.conferencemanagementsystem.app.conference.service;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import uvt.tw.conferencemanagementsystem.api.dto.conference.ConferenceRequestDto;
 import uvt.tw.conferencemanagementsystem.api.dto.conference.ConferenceResponseDto;
+import uvt.tw.conferencemanagementsystem.api.dto.conference.ConferenceStatusUpdateDto;
 import uvt.tw.conferencemanagementsystem.api.dto.conference.enums.ConferenceStatus;
 import uvt.tw.conferencemanagementsystem.api.dto.user.enums.UserRole;
 import uvt.tw.conferencemanagementsystem.api.exception.ConferenceNotFoundException;
 import uvt.tw.conferencemanagementsystem.api.exception.OwnershipException;
+import uvt.tw.conferencemanagementsystem.api.exception.StatusNotValidException;
 import uvt.tw.conferencemanagementsystem.api.exception.UserNotFoundException;
 import uvt.tw.conferencemanagementsystem.app.conference.model.ConferenceEntity;
 import uvt.tw.conferencemanagementsystem.app.conference.repository.ConferenceRepository;
@@ -99,7 +106,57 @@ public class ConferenceService {
                 () ->
                     new ConferenceNotFoundException(
                         String.format("Conference with id: %d not found", conferenceId)));
+    validateConferenceOwnership(conference);
     conferenceRepository.deleteById(conference.getId());
+  }
+
+  @Transactional
+  public ConferenceResponseDto updateConferenceStatus(
+      ConferenceStatusUpdateDto statusDto, Long conferenceId) {
+    ConferenceEntity conference =
+        conferenceRepository
+            .findById(conferenceId)
+            .orElseThrow(
+                () ->
+                    new ConferenceNotFoundException(
+                        String.format("Conference with id: %d not found", conferenceId)));
+
+    validateConferenceOwnership(conference);
+    if (!getValidStatuses(conference.getStatus()).contains(statusDto.getConferenceStatus())) {
+      throw new StatusNotValidException(
+          String.format(
+              "Conference with status {%s} cannot be moved to status {%s}",
+              conference.getStatus(), statusDto.getConferenceStatus()));
+    }
+
+    conference.setStatus(statusDto.getConferenceStatus());
+    conference = conferenceRepository.saveAndFlush(conference);
+
+    return ConferenceConverter.convertToResponseDto(conference);
+  }
+
+  @Transactional
+  @Async
+  @Scheduled(cron = "0 0 8 * * ?")
+  public void startScheduledConferences() {
+    ArrayList<ConferenceEntity> startingConferences =
+        (ArrayList<ConferenceEntity>)
+            conferenceRepository.getConferenceEntitiesByStatusInAndStartDate(
+                List.of(ConferenceStatus.PUBLISHED), LocalDate.now());
+    startingConferences.forEach(c -> c.setStatus(ConferenceStatus.ONGOING));
+    conferenceRepository.saveAllAndFlush(startingConferences);
+  }
+
+  @Transactional
+  @Async
+  @Scheduled(cron = "0 0 22 * * ?")
+  public void endScheduledConferences() {
+    ArrayList<ConferenceEntity> endingConferences =
+        (ArrayList<ConferenceEntity>)
+            conferenceRepository.getConferenceEntitiesByStatusInAndEndDate(
+                List.of(ConferenceStatus.ONGOING), LocalDate.now());
+    endingConferences.forEach(c -> c.setStatus(ConferenceStatus.ONGOING));
+    conferenceRepository.saveAllAndFlush(endingConferences);
   }
 
   public void validateConferenceOwnership(ConferenceEntity conference) {
@@ -122,5 +179,17 @@ public class ConferenceService {
     entity.setMaxAttendees(requestDto.getMaxAttendees());
     entity.setRegistrationDeadline(requestDto.getRegistrationDeadline());
     entity.setWebsiteUrl(requestDto.getWebsiteUrl());
+  }
+
+  private List<ConferenceStatus> getValidStatuses(ConferenceStatus conferenceStatus) {
+    if (Objects.requireNonNull(conferenceStatus) == ConferenceStatus.DRAFT) {
+      return List.of(ConferenceStatus.PUBLISHED);
+    }
+    return List.of(
+        ConferenceStatus.DRAFT,
+        ConferenceStatus.CANCELLED,
+        ConferenceStatus.COMPLETED,
+        ConferenceStatus.ONGOING,
+        ConferenceStatus.PUBLISHED);
   }
 }
